fixed bug with passwords not appearing after logging in, refactored axios calls into their own seperate file
