+++ frontend/src/components/Login/Login.js
@@ -1,17 +1,18 @@
 import { useState } from 'react'
-import Avatar from '@material-ui/core/Avatar';
-import Button from '@material-ui/core/Button';
-import CssBaseline from '@material-ui/core/CssBaseline';
-import TextField from '@material-ui/core/TextField';
-import Link from '@material-ui/core/Link';
-import Paper from '@material-ui/core/Paper';
-import Box from '@material-ui/core/Box';
-import Grid from '@material-ui/core/Grid';
+import Avatar from '@material-ui/core/Avatar'
+import Button from '@material-ui/core/Button'
+import CssBaseline from '@material-ui/core/CssBaseline'
+import TextField from '@material-ui/core/TextField'
+import Link from '@material-ui/core/Link'
+import Paper from '@material-ui/core/Paper'
+import Box from '@material-ui/core/Box'
+import Grid from '@material-ui/core/Grid'
+import { loginUser } from '../axiosCalls.js'
 import axios from 'axios'
-import Typography from '@material-ui/core/Typography';
-import { makeStyles } from '@material-ui/core/styles';
-import { toast } from 'react-toastify';
-import 'react-toastify/dist/ReactToastify.css';
+import Typography from '@material-ui/core/Typography'
+import { makeStyles } from '@material-ui/core/styles'
+import { toast } from 'react-toastify'
+import 'react-toastify/dist/ReactToastify.css'
 
 function Copyright() {
   return (
@@ -62,7 +63,7 @@ function SignIn({ setCurrentUserUuid }) {
   const [email, setEmail] = useState('')
   const [password, setPassword] = useState('')
 
-  const loginUser = function (event, setCurrentUserUuid) {
+  const loginUser = function (event, setCurrentUserUuid, email, password ) {
     event.preventDefault()
 
     axios.post("http://localhost:8080/login", { email, password })
@@ -121,7 +122,7 @@ function SignIn({ setCurrentUserUuid }) {
               variant="contained"
               color="primary"
               className={classes.submit}
-              onClick={(event) => loginUser(event, setCurrentUserUuid)}
+              onClick={(event) => loginUser(event, setCurrentUserUuid, email, password )}
             >
               Sign In
             </Button>
