+++ frontend/src/components/PasswordFormGenerate/PasswordForm.js
@@ -1,8 +1,7 @@
 import { React, useState } from 'react'
 import { makeStyles } from '@material-ui/core/styles'
-import axios from 'axios'
-import { toast } from 'react-toastify';
-import 'react-toastify/dist/ReactToastify.css';
+import { saveNewPasswrodForm } from '../axiosCalls.js'
+import 'react-toastify/dist/ReactToastify.css'
 import {
   Button,
   Select,
@@ -14,7 +13,6 @@ import {
   FormControlLabel,
   MenuItem
 } from '@material-ui/core'
-let generator = require('generate-password')
 
 const categories = [
   "Social",
@@ -39,34 +37,6 @@ const useStyles = makeStyles((theme) => ({
     width: '750px'
   }
 }));
-
-const saveNewPasswrod = function (event, sessionUuid, category, url, sliderValue, checked) {
-
-  if (!checked.numbers && !checked.symbols && !checked.lowercas && !checked.uppercase) {
-    return toast.error("You can't leave all checkboxes empty when making a password!")
-  }
-
-  const password = generator.generateMultiple(3, {
-    length: 20,
-    uppercase: checked.uppercase,
-    lowercase: checked.lowercase,
-    numbers: checked.numbers,
-    symbols: checked.symbols
-  });
-
-  axios.post("http://localhost:8080/passwords/create", { sessionUuid, passwordText: password[0], category, url })
-  .then((response) => {
-    if (response) {
-      console.log(response.data)
-      toast.success(response.data)
-    }
-  }).catch((error) => {
-    if (error) {
-      console.log(error.response.data.message)
-      toast.error(error.response.data.message)
-    }
-  })
-}
 
 function PasswordForm({ sessionUuid }) {
   const classes = useStyles();
@@ -162,7 +132,7 @@ function PasswordForm({ sessionUuid }) {
             />
           </Grid>
         </Grid>
-        <Button onClick={(event) => saveNewPasswrod(event, sessionUuid, category, url, sliderValue, checked)} >Generate Password</Button>
+        <Button onClick={(event) => saveNewPasswrodForm(event, sessionUuid, category, url, sliderValue, checked)} >Generate Password</Button>
       </FormControl>
     </div>
   )
